add pie-chart and rename files
