+++ src/index.js
@@ -1,7 +1,7 @@
 import "./index.scss";
 
 import { AppComponent } from "./components/app/app";
-import { PieChartComponent } from "./components/pieChart/pieChart";
+import { PieChartComponent } from "./components/pie-chart/pie-chart";
 
 const root = document.querySelector("#root");
 new AppComponent(root).mount();
