+++ src/index.scss
@@ -4,4 +4,4 @@
 
 @import "./components/page/page";
 @import "./components/app/app";
-@import "./components/pieChart/pieChart";
+@import "./components/pie-chart/pie-chart";
